requirements.txt
aiohttp==3.7.4.post0
anytree==2.8.0
apipkg==1.5
apparmor==3.0.1
appdirs==1.4.4
application-utility==1.3.2
async-timeout==3.0.1
attrs==21.2.0
bcrypt==3.2.0
Beaker==1.11.0
beautifulsoup4==4.9.3
breezy==3.2.0
btrfsutil==5.13
CacheControl==0.12.6
cachetools==4.2.2
caffeine-ng==3.5.1
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
chrome-gnome-shell==0.0.0
colorama==0.4.4
configobj==5.1.0.dev0
contextlib2==0.6.0.post1
cryptography==3.4.7
Cython==0.29.24
dateparser==1.0.0
devtools==0.6.1
DictObject==1.0.5
distlib==0.3.2
distro==1.5.0
dnspython==1.16.0
docopt==0.6.2
duplicity==0.8.19
evdev==1.4.0
ewmh==0.1.6
fasteners==0.16
filelock==3.0.12
future==0.18.2
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gufw==21.4.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.19.1
idna==3.2
importlib-metadata==4.6.1
iniconfig==1.1.1
layoutswitcherlib==0.8.22
LibAppArmor==3.0.1
lit==12.0.1.dev0
louis==3.18.0
luckydonald-utils==0.76
lutris==0.5.8.4
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
meson==0.59.0
monotonic==1.6
more-itertools==8.8.0
msgpack==1.0.2
multidict==5.1.0
netsnmp-python==1.0a1
npyscreen==4.10.5
numpy==1.20.3
oauth2client==4.1.3
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.21.5
paramiko==2.7.2
patiencediff==0.2.2
pep517==0.11.0
Pillow==8.3.1
pluggy==0.13.1
ply==3.11
progress==1.5
protobuf==3.17.3
psutil==5.8.0
pulsectl==21.5.18
pwquality==1.4.4
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.7.1
pycairo==1.20.0
pycparser==2.20
PyDrive==1.3.1
Pygments==2.9.0
PyGObject==3.40.1
PyNaCl==1.4.0
PyOpenGL==3.1.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-sip==12.9.0
pysmbc==1.0.23
pytest==6.2.4
pytest-mock==3.1.1
pytest-runner==5.3.1
pytg==0.4.10
python-binance==1.0.12
python-dateutil==2.8.2
python-xlib==0.31
pytz==2021.1
pyxdg==0.27
PyYAML==5.4.1
ranger-fm==1.9.3
regex==2021.8.3
reportlab==3.5.68
requests==2.26.0
requests-file==1.5.1
resolvelib==0.5.5
retrying==1.3.3
rsa==4.7.2
setproctitle==1.2.2
six==1.16.0
soupsieve==2.2.1
TA-Lib==0.4.21
team==1.0
tldextract==3.1.0
toml==0.10.2
tomli==1.2.0
typing-extensions==3.10.0.0
tzlocal==2.1
ueberzug==18.1.9
ufw==0.36
ujson==4.0.2
uritemplate==3.0.1
urllib3==1.26.6
urwid==2.1.2
validate==5.1.0.dev0
webencodings==0.5.1
websocket-client==1.1.1
websockets==9.1
yarl==1.6.3
youtube-dl==2021.6.6
zipp==3.5.0
